backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings>=2.0.0
psycopg2-binary
sqlalchemy
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
